tests/requirements.txt
pytest>=4.0
pytest-cov>=2.6.1
flywheel-sdk>=10.4